Bluetooth: Setting authentication req to Low for Pointing devices.
If a HID Mouse is 2.0 device, it doesn't support bonding procedure.
If it is a 2.1 device, then authentication is mandatory and the
security level would be reset to HIGH. Since authentication is now
not performed for 2.0 HID mouse devices, SDP is triggerred before
L2CAP Control Channel creation which updates upper layer of pairing
being complete.
CRs-Fixed: 465421

Change-Id: I8568e61b24402d1233000a6d13cd33b782fb974a