Bluetooth: Add support for disconnecting HID based on priority
This patch adds support for disconnecting the incoming HID connected based
on priority of remote HID Device. If user does not want to connect HID
profile for a particular remote device, priority of remote device would be
set to 0. Based on priority of remote device, incoming connection would be
disconnected if priority is 0.

CRs-Fixed: 532234

Conflicts:

	btif/include/btif_storage.h
	btif/src/btif_storage.c
Change-Id: I97002a46a16d386a6c178445ac825788d6bf40c9