Bluetooth: To prevent LE unpair if encryption fails to start
This patch prevents unapiring the LE device if the encryption
fails to start as the link on which the encryption is supposed
to start, is already disconnected.

Change-Id: I7fc74580eb1569dcf17b397069b79045791085e9
CRs-Fixed: 685521

Bluetooth: Change to prevent dev unpair caused by host disconn

This patch includes changes to prevent unpairing the device
in case a disconnection is initiated by the host, when an encryption
is still in progress.

CRs-Fixed: 694033
Change-Id: I134fa0f1f7f574007cce4814ff8f677e747221ab