HID: Fix IOT issue with Motorolla KZ500 Keyboard
This patch fixes IOT issue with Motorolla KZ500 Keyboard, which claims
support for multiouch functionality in SDP repord descriptor, but this
functionality doesn't work because of issues in usb hid kernel driver.
To handle this keyboard, digitizer record in report descriptor is rmeoved
and the remaining descriptor is passed to kernel usb hid driver during
connection, which interprets the keyboard as generic keyboard instead of
multitouch keyboard and normal functioning of keyboard works fine thereafter.

Change-Id: I888b4a8997296ebbf3d9d23dd1c52ade230f984e